LPD-16079 gradle-plugins-workspace: WorkspaceExtension: only retrieves the product info once in the constructor
The reason that workspace has been slow for a while is that the DownloadCommand
for the .product_info.json file was being run every time a product info was
queried.

This was higlighted by the ISP outage in Diamond Bar during Jan 2024. Since
there are many queries (about 6 when WorkspaceExtension is initialized and then
one per project in the workspace), and each could take up to ten seconds
(5 second timout x2), this resulted in the workspace being essentially unable
to progress past the configuration phase.

This commit just performs the downloads once, for a maximum timeout of 10
seconds.