LPD-29543 Remove the destination reference in AntivirusAsyncStatisticsManager
# breaking

## What modules/dxp/apps/antivirus/antivirus-async-store/src/main/java/com/liferay/antivirus/async/store/jmx/AntivirusAsyncStatisticsManager.java

The OSGi declarative reference to Destination in the constructor parameter is removed. The destination is now obtained from the MessageBus service.

## Why

Avoid referencing Destination in preparation for the destination registration refactor.

## Alternative

No action needs to be taken, as this class should never be instantiated manually, as it's managed by the OSGi framework.

----