LPD-38883 Avoid sending duplicated values when submitting the form
Before we had a condition to return null instead of a hidden input when the locale was the same as the editingLanguageId, avoiding the value to be submitted twice.

This condition was necessary because the current inputs the user is editing already have the values on the screen (and also on the DOM), thus there isn't a need for hidden values for the current editingLanguageId. When this condition was removed, values started to be sent twice.

Since the web content translation manager now needs all the hidden inputs to calculate their translations, the condition was changed to not add the name attribute into the input hidden that has the same locale as the current editingLanguageId; this way, the value from the input hidden won't be submitted and the value will be sent only once.